requirements.txt
fastapi>=0.68.0,<0.69.0
sqlalchemy
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
psycopg2
pandas
python-dotenv